+++ cucumber-guice/src/main/java/io/cucumber/guice/GuiceFactory.java
@@ -10,6 +10,8 @@ import org.apiguardian.api.API;
 import java.util.Collection;
 import java.util.HashSet;
 
+import static io.cucumber.guice.InjectorSourceFactory.createDefaultScenarioModuleInjectorSource;
+import static io.cucumber.guice.InjectorSourceFactory.instantiateUserSpecifiedInjectorSource;
 import static io.cucumber.guice.InjectorSourceFactory.loadInjectorSourceFromProperties;
 import static java.lang.String.format;
 
@@ -29,6 +31,10 @@ public final class GuiceFactory implements ObjectFactory {
 
     public GuiceFactory() {
         this.injectorSourceFromProperty = loadInjectorSourceFromProperties(CucumberProperties.create());
+        // Eager init to allow for static binding prior to before all hooks
+        if (this.injectorSourceFromProperty != null) {
+            injector = instantiateUserSpecifiedInjectorSource(this.injectorSourceFromProperty).getInjector();
+        }
     }
 
     @Override
@@ -40,6 +46,9 @@ public final class GuiceFactory implements ObjectFactory {
             if (hasInjectorSource(stepClass)) {
                 checkOnlyOneClassHasInjectorSource(stepClass);
                 withInjectorSource = stepClass;
+                // Eager init to allow for static binding prior to before all
+                // hooks
+                injector = instantiateUserSpecifiedInjectorSource(withInjectorSource).getInjector();
             }
         }
         stepClasses.add(stepClass);
@@ -69,9 +78,10 @@ public final class GuiceFactory implements ObjectFactory {
     }
 
     public void start() {
+        // Last minute init. Neither properties not annotations provided an
+        // injector source.
         if (injector == null) {
-            injector = new InjectorSourceFactory(withInjectorSource).create()
-                    .getInjector();
+            injector = createDefaultScenarioModuleInjectorSource().getInjector();
         }
         scenarioScope = injector.getInstance(ScenarioScope.class);
         scenarioScope.enterScope();